fix some HLSL errors fix a warning on some ft2 code
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@10597 d7cf8633-e32d-0410-b094-e92efae38249

+++ ft2.c
@@ -963,7 +963,7 @@ void Font_UnloadFont(ft2_font_t *font)
 
 	if (font->attachments && font->attachmentcount)
 	{
-		for (i = 0; i < font->attachmentcount; ++i) {
+		for (i = 0; i < (int)font->attachmentcount; ++i) {
 			if (font->attachments[i].data)
 				Mem_Free(font->attachments[i].data);
 		}
